pyproject.toml
[project]
name = "scoliosis-segmentation-ms"
version = "0.1.0"
description = "Microservicio Python para servir el pipeline de segmentacion de escoliosis."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "pillow>=10.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
ml = [
    "numpy>=1.26.0",
    "torch>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

